!!! fixed bug in training code
This is a serious and unfortunate error, caused by a one-line problem.
In this case PyTorch does automatically detect parameters that are
properties of a class and return them correctly from Model.params()
Unfortunately it doesn't look into lists. So you need to use a special
kind of list if you are going to dynamically create lists of layers.

The upshot of this is that any training that used a deep model was
only training the last layer. This also meant that when checkpoints
get loaded the lower layers remained uninitalised. How is it possible
that we actually saw learning and an increase in performance?
It seems that for Goofspiel 13, just training the biases helped
a lot and was enough to beat a random player. The freshly initalised
layers were probably approximately normally distributed with mean zero
and so just added some extra variance to the answers.

There are two intesting lessons to take away from this, as always
machine learning systems often mask bugs in their implementation,
background non-explicit magic is evil and will bite you eventually.

AK